[FIX] 게시글 상세 조회 시 멘토 정보 및 리뷰 정보들 추가

+++ src/main/java/com/preview/preview/module/review/application/dto/ReviewCreateRequestDto.java
@@ -6,5 +6,5 @@ import lombok.*;
 @Setter
 public class ReviewCreateRequestDto {
     private String contents;
-    private Integer grade;
+    private float grade;
 }
